Split test_multiprocessing and enabled cluster tests
