+++ python/ray/tests/BUILD
@@ -103,6 +103,8 @@ py_test_module_list(
     "test_multinode_failures.py",
     "test_multinode_failures_2.py",
     "test_multiprocessing.py",
+    "test_multiprocessing_2.py",
+    "test_multiprocessing_3.py",
     "test_object_assign_owner.py",
     "test_placement_group.py",
     "test_placement_group_2.py",
@@ -372,6 +374,8 @@ py_test_module_list(
     "test_basic_5.py",
     "test_asyncio.py",
     "test_multiprocessing.py",
+    "test_multiprocessing_2.py",
+    "test_multiprocessing_3.py",
     "test_list_actors.py",
     "test_list_actors_2.py",
     "test_list_actors_3.py",
